Use mkdtempSync from Node
Previously we hand-wrote some code to create a temporary directory
and clean it up after parsing the package set. However, this could
fail (if, for example, the process was cancelled by the user). In
these cases the directory would be left intact and subsequent runs
of the integration test would fail.

Using Node to access the operating system's builtin handling for
temporary directories is more robust and predictable.